+++ Changelog.md
@@ -1,11 +1,13 @@
 ## OpenStudio-HPXML v1.2.0 (Pending)
 
 __New Features__
-- Allow `Slab/ExposedPerimeter` to be zero.
+- **Breaking change**: Heating/cooling component loads no longer calculated by default for faster performance; use `--add-component-loads` argument if desired.
 - **Breaking change**: Replaces `Site/extension/ShelterCoefficient` with `Site/ShieldingofHome`.
+- Allows `Slab/ExposedPerimeter` to be zero.
 - Removes `ClothesDryer/ControlType` from being a required input, it is not used.
 - Moves additional error-checking from the ruby measure to the schematron validator.
 - Adds more detail to error messages regarding the wrong data type in the HPXML file.
+- Adds error-checking for negative SEEReq results for shared cooling systems.
 - Relaxes tolerance for duct leakage to outside warning when ducts solely in conditioned space.
 
 __Bugfixes__